+++ src/pharmpy/datainfo.py
@@ -35,3 +35,19 @@ class DataInfo:
     @id_label.setter
     def id_label(self, name):
         self._set_one_label_to_type(name, 'id')
+
+    @property
+    def dv_label(self):
+        return self._get_one_label_by_type('dv')
+
+    @dv_label.setter
+    def dv_label(self, name):
+        self._set_one_label_to_type(name, 'dv')
+
+    @property
+    def idv_label(self):
+        return self._get_one_label_by_type('idv')
+
+    @idv_label.setter
+    def idv_label(self, name):
+        self._set_one_label_to_type(name, 'idv')