+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+next version
+------------
+
+* Add modeling.get_individuals
+
 0.40.0 (2021-12-16)
 -------------------